+++ tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use Nexxai\FreeTsa\Tests\TestCase;
+use Nexxai\Rfc3161\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);